fix: rename plugin-inputs to plugin-input and update related imports
