+++ src/types/index.ts
@@ -1,4 +1,4 @@
 export * from "./permits";
 export * from "./context";
 export * from "./env";
-export * from "./plugin-inputs";
+export * from "./plugin-input";
